requirements.txt
lxml==6.0.2
pillow==11.3.0
python-pptx==1.0.2
typing_extensions==4.15.0
xlsxwriter==3.2.9
